Refactor reducer code
Move to separate files for easier management in the future.
