+++ app/reducers/reducers.js
@@ -1,0 +1,12 @@
+'use strict';
+
+import { combineReducers } from 'redux';
+import authState from './auth';
+import tripsState from './trips';
+
+const appReducers = combineReducers({
+  authState,
+  tripsState
+});
+
+export default appReducers;
