+++ app/reducers/trips.js
@@ -1,19 +1,7 @@
 'use strict';
 
-import { combineReducers } from 'redux';
 import _ from 'underscore';
 import {
-  LOGIN_SAVE_EMAIL,
-  LOGIN_SAVE_PASSWORD,
-  SIGNUP_SAVE_NAME,
-  SIGNUP_SAVE_EMAIL,
-  SIGNUP_SAVE_PASSWORD,
-  API_LOGIN_REQUEST,
-  API_LOGIN_SUCCESS,
-  API_LOGIN_FAILURE,
-  API_SIGNUP_REQUEST,
-  API_SIGNUP_SUCCESS,
-  API_SIGNUP_FAILURE,
   CREATE_TRIP_SAVE_TITLE,
   CREATE_TRIP_SAVE_DEST,
   CREATE_TRIP_SAVE_VISIBILITY,
@@ -28,61 +16,10 @@ import {
   API_CREATE_TRIP_SUCCESS,
   API_CREATE_TRIP_FAILURE,
   LOGOUT
-} from './actions/actions';
-import { initialAuthState, initialTripsState } from './constants';
+} from '../actions/actions';
+import { initialTripsState } from '../constants';
 
-function authState(state = initialAuthState, action) {
-  switch (action.type) {
-    case LOGIN_SAVE_EMAIL:
-      return { ...state, email: action.email };
-    case LOGIN_SAVE_PASSWORD:
-      return { ...state, password: action.password };
-    case SIGNUP_SAVE_NAME:
-      return { ...state, newName: action.name };
-    case SIGNUP_SAVE_EMAIL:
-      return { ...state, newEmail: action.email };
-    case SIGNUP_SAVE_PASSWORD:
-      return { ...state, newPassword: action.password };
-    case API_LOGIN_REQUEST:
-    case API_SIGNUP_REQUEST:
-      return {
-        ...(_.omit(state, 'error')),
-        isFetching: true
-      };
-    case API_LOGIN_SUCCESS:
-      return {
-        ...(_.omit(state, ['error', 'email', 'password'])),
-        isFetching: false,
-        user: action.user,
-        token: action.token
-      };
-    case API_SIGNUP_SUCCESS:
-      return {
-        ...(_.omit(state, [
-          'error',
-          'newName',
-          'newEmail',
-          'newPassword'
-        ])),
-        isFetching: false,
-        user: action.user,
-        token: action.token
-      };
-    case API_LOGIN_FAILURE:
-    case API_SIGNUP_FAILURE:
-      return {
-        ...state,
-        isFetching: false,
-        error: action.error
-      };
-    case LOGOUT:
-      return initialAuthState;
-  }
-
-  return state;
-}
-
-function tripsState(state = initialTripsState, action) {
+export default function tripsState(state = initialTripsState, action) {
   switch (action.type) {
     case CREATE_TRIP_SAVE_TITLE:
       return {
@@ -150,10 +87,3 @@ function tripsState(state = initialTripsState, action) {
 
   return state;
 }
-
-const appReducers = combineReducers({
-  authState,
-  tripsState
-});
-
-export default appReducers;
